Rust cleanup
