+++ nix-rust/src/error.rs
@@ -1,0 +1,5 @@
+#[derive(Debug)]
+pub enum Error {
+    Misc(String),
+    Foreign(libc::c_void), // == std::exception_ptr
+}
